// server/admin/frontend/src/generated-client/models/CreateEntryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateEntryResponse
 */
export interface CreateEntryResponse {
    /**
     * The path of the created entry
     * @type {string}
     * @memberof CreateEntryResponse
     */
    path: string;
}

/**
 * Check if a given object implements the CreateEntryResponse interface.
 */
export function instanceOfCreateEntryResponse(value: object): value is CreateEntryResponse {
    if (!('path' in value) || value['path'] === undefined) return false;
    return true;
}

export function CreateEntryResponseFromJSON(json: any): CreateEntryResponse {
    return CreateEntryResponseFromJSONTyped(json, false);
}

export function CreateEntryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateEntryResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'path': json['path'],
    };
}

export function CreateEntryResponseToJSON(json: any): CreateEntryResponse {
    return CreateEntryResponseToJSONTyped(json, false);
}

export function CreateEntryResponseToJSONTyped(value?: CreateEntryResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'path': value['path'],
    };
}